# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A REST API for managing projects",
        "title": "Prodyo Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/bugs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all bugs for a specific task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bugs"
                ],
                "summary": "Get all bugs",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bugs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Bug"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid task_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve bugs",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new bug for a task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bugs"
                ],
                "summary": "Create a new bug",
                "parameters": [
                    {
                        "description": "Bug data",
                        "name": "bug",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateBugRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Bug created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to create bug",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bugs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific bug by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bugs"
                ],
                "summary": "Get bug by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Bug ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bug details",
                        "schema": {
                            "$ref": "#/definitions/models.Bug"
                        }
                    },
                    "400": {
                        "description": "Invalid bug ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Bug not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/improvements": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all improvements for a specific task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "improvements"
                ],
                "summary": "Get all improvements",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of improvements",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Improv"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid task_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve improvements",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new improvement for a task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "improvements"
                ],
                "summary": "Create a new improvement",
                "parameters": [
                    {
                        "description": "Improvement data",
                        "name": "improvement",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateImprovRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Improvement created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to create improvement",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/improvements/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific improvement by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "improvements"
                ],
                "summary": "Get improvement by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Improvement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Improvement details",
                        "schema": {
                            "$ref": "#/definitions/models.Improv"
                        }
                    },
                    "400": {
                        "description": "Invalid improvement ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Improvement not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateBugRequest": {
            "type": "object",
            "properties": {
                "assignee_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "task_id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.CreateImprovRequest": {
            "type": "object",
            "properties": {
                "assignee_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "task_id": {
                    "type": "string"
                }
            }
        },
...
        "models.Bug": {
            "type": "object",
            "properties": {
                "assignee": {
                    "$ref": "#/definitions/models.User"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "task_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.Improv": {
            "type": "object",
            "properties": {
                "assignee": {
                    "$ref": "#/definitions/models.User"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "task_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}